🧹 fix: integer counts + targeted token_count migration warning
Two findings from the comprehensive review:

1. `remaining_tokens` and `messages_to_refine` are token/message counts
   and are always integers in the runtime (`Number.isFinite(...)` guards
   already assume integer semantics). `z.number().positive()` accepted
   fractional values like `2.5`, which was semantically confusing and
   would round oddly against the runtime's `>=` / `<=` comparisons. Add
   `.int()` to both count-based branches; `token_ratio` stays fractional.

2. Anyone upgrading with `trigger.type: 'token_count'` in their YAML got
   the generic "Invalid summarization config" warning plus a flattened
   Zod error. Detect that specific case in `loadSummarizationConfig` and
   emit a migration-friendly message that names the three valid
   replacements. Export the function so the behavior is unit-testable.

Also added a parameterized passthrough test covering `remaining_tokens`
and `messages_to_refine` shaping, complementing the existing
`token_ratio` coverage.

+++ packages/data-provider/src/config.ts
@@ -1027,11 +1027,11 @@ export const summarizationTriggerSchema = z.discriminatedUnion('type', [
   }),
   z.object({
     type: z.literal('remaining_tokens'),
-    value: z.number().finite().positive(),
+    value: z.number().finite().int().positive(),
   }),
   z.object({
     type: z.literal('messages_to_refine'),
-    value: z.number().finite().positive(),
+    value: z.number().finite().int().positive(),
   }),
 ]);
